+++ tests/integration/database_model_orm/__init__.py
@@ -1,4 +1,4 @@
 # pylint: disable=missing-module-docstring
 __all__ = [
-    'conftest',
+        'test_int__postgres_orm_models',
 ]